energy bar

+++ TP8-DVJ2/Assets/Scripts/Gameplay/Ship.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public abstract class Ship : MonoBehaviour
 {
-    protected int Energy;
+    protected float Energy;
 
     public abstract void Shoot();
     public abstract void Move();
